IoT/requirements.txt
adafruit_circuitpython_ble==8.3.8
ecdsa==0.18.0
logtail_python==0.1.3
pycryptodome==3.15.0
pysftp==0.2.9
